end of evaluator module
